refactor: merge bot and worker classes and ensure testability of the discordclient
