+++ NLBE-Bot/Interfaces/ICommandEventHandler.cs
@@ -1,0 +1,8 @@
+namespace NLBE_Bot.Interfaces;
+
+using DSharpPlus.CommandsNext;
+
+internal interface ICommandEventHandler
+{
+	public void Register(ICommandsNextExtension commands);
+}
