+++ NLBE-Bot/Interfaces/IGuildMemberEventHandler.cs
@@ -1,0 +1,6 @@
+namespace NLBE_Bot.Interfaces;
+
+internal interface IGuildMemberEventHandler
+{
+	public void Register(IDiscordClientWrapper client);
+}
