+++ NLBE-Bot/Interfaces/IMessageEventHandler.cs
@@ -1,0 +1,6 @@
+namespace NLBE_Bot.Interfaces;
+
+internal interface IMessageEventHandler
+{
+	public void Register(IDiscordClientWrapper client);
+}
