+++ NLBE-Bot/Interfaces/ITournamentService.cs
@@ -2,6 +2,8 @@ namespace NLBE_Bot.Interfaces;
 
 using DSharpPlus.Entities;
 using FMWOTB.Tournament;
+using NLBE_Bot.Models;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -12,4 +14,12 @@ internal interface ITournamentService
 	public Task<List<WGTournament>> InitialiseTournaments(bool all);
 
 	public Task ShowTournamentInfo(DiscordChannel channel, WGTournament tournament, string titel);
+
+	public Task<List<Tier>> ReadTeams(DiscordChannel channel, DiscordMember member, string guildName, string[] parameters_as_in_hoeveelste_team);
+
+	public Task<List<Tuple<ulong, string>>> GetIndividualParticipants(List<Tier> teams, DiscordGuild guild);
+
+	public bool CheckIfAllWithinRange(string[] tiers, int min, int max);
+
+	public Task<List<string>> GetMentions(List<Tuple<ulong, string>> memberList, ulong guildID);
 }
