--- NLBE-Bot/Services/WeeklyEventService.cs
@@ -1,10 +1,0 @@
-namespace NLBE_Bot.Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-internal class WeeklyEventService
-{
-}
